Test with Java 21 and Java 17
Java 21 released Sep 19, 2023. Full support for Java 21 was announced
in November 2023.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

This does not test Java 11 because the Java 17 build and the Java 21
build are both generating Java 11 byte code.  We've detected no issues
related to building Java 11 byte code with Java 17 and Java 21.